MessageList changes: - Use long instead of int for edition numbers - Use MessageManager function instead of plain db4o query...

+++ src/plugins/Freetalk/OwnMessageList.java
@@ -27,7 +27,7 @@ public abstract class OwnMessageList extends MessageList {
 
 	}
 
-	public OwnMessageList(OwnIdentity newAuthor, int newIndex) {
+	public OwnMessageList(OwnIdentity newAuthor, long newIndex) {
 		super(newAuthor, newIndex);
 	}
